Сделано

+++ task.30.10/2/task6/task6.cpp
@@ -3,11 +3,11 @@
 using namespace std;
 
 int main() {
-    double a, b, c;
+    int a, b, c;
     cout << "Введите цены: " << endl;
     cin >> a >> b >> c;
 
-    double d = max({a, b, c});
+    int d = max({a, b, c});
     cout << "Стоимость телефона: " << d << endl;
     return 0;
 }
